factor out service account secret & key management
