+++ styx-scheduler-service/src/main/java/com/spotify/styx/docker/DockerRunner.java
@@ -117,8 +117,10 @@ public interface DockerRunner extends Closeable {
 
   static DockerRunner kubernetes(NamespacedKubernetesClient kubernetesClient, StateManager stateManager,
       Stats stats, ServiceAccountKeyManager serviceAccountKeyManager) {
+    final KubernetesGCPServiceAccountSecretManager serviceAccountSecretManager =
+        new KubernetesGCPServiceAccountSecretManager(kubernetesClient, serviceAccountKeyManager);
     final KubernetesDockerRunner dockerRunner =
-        new KubernetesDockerRunner(kubernetesClient, stateManager, stats, serviceAccountKeyManager);
+        new KubernetesDockerRunner(kubernetesClient, stateManager, stats, serviceAccountSecretManager);
 
     dockerRunner.init();
