+++ styx-scheduler-service/src/main/java/com/spotify/styx/docker/KubernetesDockerRunner.java
@@ -23,7 +23,6 @@ package com.spotify.styx.docker;
 import static com.spotify.styx.docker.KubernetesPodEventTranslator.translate;
 import static com.spotify.styx.serialization.Json.OBJECT_MAPPER;
 import static com.spotify.styx.state.RunState.State.RUNNING;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.stream.Collectors.toSet;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -32,14 +31,9 @@ import com.github.rholder.retry.Retryer;
 import com.github.rholder.retry.RetryerBuilder;
 import com.github.rholder.retry.StopStrategies;
 import com.github.rholder.retry.WaitStrategies;
-import com.google.api.client.googleapis.json.GoogleJsonResponseException;
-import com.google.api.services.iam.v1.model.ServiceAccountKey;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Throwables;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.hash.Hashing;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import com.spotify.styx.ServiceAccountKeyManager;
 import com.spotify.styx.model.Event;
 import com.spotify.styx.model.EventVisitor;
 import com.spotify.styx.model.ExecutionDescription;
@@ -50,7 +44,6 @@ import com.spotify.styx.state.Message;
 import com.spotify.styx.state.RunState;
 import com.spotify.styx.state.StateManager;
 import com.spotify.styx.state.Trigger;
-import com.spotify.styx.util.GcpUtil;
 import com.spotify.styx.util.TriggerUtil;
 import io.fabric8.kubernetes.api.model.ContainerBuilder;
 import io.fabric8.kubernetes.api.model.EnvVar;
@@ -60,8 +53,6 @@ import io.fabric8.kubernetes.api.model.PodBuilder;
 import io.fabric8.kubernetes.api.model.PodList;
 import io.fabric8.kubernetes.api.model.PodSpecBuilder;
 import io.fabric8.kubernetes.api.model.Secret;
-import io.fabric8.kubernetes.api.model.SecretBuilder;
-import io.fabric8.kubernetes.api.model.SecretList;
 import io.fabric8.kubernetes.api.model.SecretVolumeSource;
 import io.fabric8.kubernetes.api.model.SecretVolumeSourceBuilder;
 import io.fabric8.kubernetes.api.model.Volume;
@@ -73,11 +64,9 @@ import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.NamespacedKubernetesClient;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
+import io.norberg.automatter.AutoMatter;
 import java.io.IOException;
 import java.net.ProtocolException;
-import java.time.Clock;
-import java.time.Duration;
-import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
@@ -90,8 +79,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 /**
  * A {@link DockerRunner} implementation that submits container executions to a Kubernetes cluster.
@@ -113,13 +100,8 @@ class KubernetesDockerRunner implements DockerRunner {
   static final String TRIGGER_TYPE = "STYX_TRIGGER_TYPE";
   private static final int DEFAULT_POLL_PODS_INTERVAL_SECONDS = 60;
   static final String STYX_WORKFLOW_SA_ENV_VARIABLE = "GOOGLE_APPLICATION_CREDENTIALS";
-  private static final String STYX_WORKFLOW_SA_ID_ANNOTATION = "styx-wf-sa";
-  private static final String STYX_WORKFLOW_SA_EPOCH_ANNOTATION = "styx-wf-sa-epoch";
-  private static final String STYX_WORKFLOW_SA_JSON_KEY_NAME_ANNOTATION = "styx-wf-sa-json-key-name";
-  private static final String STYX_WORKFLOW_SA_P12_KEY_NAME_ANNOTATION = "styx-wf-sa-p12-key-name";
   static final String STYX_WORKFLOW_SA_SECRET_NAME = "styx-wf-sa-keys";
   private static final String STYX_WORKFLOW_SA_JSON_KEY = "styx-wf-sa.json";
-  private static final String STYX_WORKFLOW_SA_P12_KEY = "styx-wf-sa.p12";
   static final String STYX_WORKFLOW_SA_SECRET_MOUNT_PATH =
       "/etc/" + STYX_WORKFLOW_SA_SECRET_NAME + "/";
 
@@ -128,10 +110,6 @@ class KubernetesDockerRunner implements DockerRunner {
       .setNameFormat("k8s-scheduler-thread-%d")
       .build();
 
-  private static final Supplier<String> DEFAULT_SECRET_EPOCH_PROVIDER = KubernetesDockerRunner::utcNoonSecretEpoch;
-
-  private static final Duration SECRET_GC_GRACE_PERIOD = Duration.ofMinutes(30);
-
   private final ScheduledExecutorService executor =
       Executors.newSingleThreadScheduledExecutor(THREAD_FACTORY);
 
@@ -139,38 +117,29 @@ class KubernetesDockerRunner implements DockerRunner {
   private final StateManager stateManager;
   private final Stats stats;
   private final int pollPodsIntervalSeconds;
-  private final ServiceAccountKeyManager serviceAccountKeyManager;
-  private final Supplier<String> secretEpochProvider;
-  private final Clock clock;
+  private final KubernetesGCPServiceAccountSecretManager serviceAccountSecretManager;
 
   private Watch watch;
 
-  private final Object secretMutationLock = new Object() {};
-
   KubernetesDockerRunner(NamespacedKubernetesClient client, StateManager stateManager, Stats stats,
-      ServiceAccountKeyManager serviceAccountKeyManager, int pollPodsIntervalSeconds,
-      Supplier<String> secretEpochProvider, Clock clock) {
+      KubernetesGCPServiceAccountSecretManager serviceAccountSecretManager, int pollPodsIntervalSeconds) {
     this.stateManager = Objects.requireNonNull(stateManager);
     this.client = Objects.requireNonNull(client).inNamespace(NAMESPACE);
     this.stats = Objects.requireNonNull(stats);
     this.pollPodsIntervalSeconds = pollPodsIntervalSeconds;
-    this.serviceAccountKeyManager = Objects.requireNonNull(serviceAccountKeyManager);
-    this.secretEpochProvider = Objects.requireNonNull(secretEpochProvider);
-    this.clock = Objects.requireNonNull(clock);
+    this.serviceAccountSecretManager = Objects.requireNonNull(serviceAccountSecretManager);
   }
 
   KubernetesDockerRunner(NamespacedKubernetesClient client, StateManager stateManager, Stats stats,
-      ServiceAccountKeyManager serviceAccountKeyManager) {
-    this(client, stateManager, stats, serviceAccountKeyManager, DEFAULT_POLL_PODS_INTERVAL_SECONDS,
-        DEFAULT_SECRET_EPOCH_PROVIDER, Clock.systemUTC());
+      KubernetesGCPServiceAccountSecretManager serviceAccountSecretManager) {
+    this(client, stateManager, stats, serviceAccountSecretManager, DEFAULT_POLL_PODS_INTERVAL_SECONDS);
   }
 
   @Override
   public String start(WorkflowInstance workflowInstance, RunSpec runSpec) throws IOException {
-    final String secretEpoch = secretEpochProvider.get();
-    ensureSecrets(workflowInstance, runSpec, secretEpoch);
+    final KubernetesSecretSpec secretSpec = ensureSecrets(workflowInstance, runSpec);
     try {
-      final Pod pod = client.pods().create(createPod(workflowInstance, runSpec, secretEpoch));
+      final Pod pod = client.pods().create(createPod(workflowInstance, runSpec, secretSpec));
       return pod.getMetadata().getName();
     } catch (KubernetesClientException kce) {
       throw new IOException("Failed to create Kubernetes pod", kce);
@@ -179,78 +148,23 @@ class KubernetesDockerRunner implements DockerRunner {
 
   @Override
   public void cleanup() throws IOException {
-
-    // Enumerate all secrets currently used by pods
-    final PodList pods = client.pods().list();
-    final Set<String> activeSecrets = pods.getItems().stream()
-        .flatMap(pod -> pod.getSpec().getVolumes().stream())
-        .map(volume -> volume.getSecret().getSecretName())
-        .collect(Collectors.toSet());
-
-    // Enumerate service account secrets to delete
-    final String epoch = secretEpochProvider.get();
-    final Instant creationDeadline = clock.instant().minus(SECRET_GC_GRACE_PERIOD);
-    final SecretList secrets = client.secrets().list();
-    final List<Secret> inactiveServiceAccountSecrets = secrets.getItems().stream()
-        // Only include service account secrets
-        .filter(secret -> secret.getMetadata().getName().startsWith(STYX_WORKFLOW_SA_SECRET_NAME))
-        // Exclude secrets in the current epoch
-        .filter(secret -> !epoch.equals(secret.getMetadata().getAnnotations().get(STYX_WORKFLOW_SA_EPOCH_ANNOTATION)))
-        // Exclude recently created secrets to mitigate races with secret creation around epoch switch
-        .filter(secret -> Instant.parse(secret.getMetadata().getCreationTimestamp()).isBefore(creationDeadline))
-        // Exclude secrets currently in use by pods
-        .filter(secret -> !activeSecrets.contains(secret.getMetadata().getName()))
-        .collect(Collectors.toList());
-
-    // Delete keys and secrets for all inactive service accounts and let them be recreated by future executions
-    for (Secret secret : inactiveServiceAccountSecrets) {
-      final String name = secret.getMetadata().getName();
-      final String serviceAcount = secret.getMetadata().getAnnotations().get(STYX_WORKFLOW_SA_ID_ANNOTATION);
-
-      try {
-        final Map<String, String> annotations = secret.getMetadata().getAnnotations();
-        final String jsonKeyName = annotations.get(STYX_WORKFLOW_SA_JSON_KEY_NAME_ANNOTATION);
-        final String p12KeyName = annotations.get(STYX_WORKFLOW_SA_P12_KEY_NAME_ANNOTATION);
-
-        LOG.info("[AUDIT] Deleting service account key: {}", jsonKeyName);
-        tryDeleteServiceAccountKey(jsonKeyName);
-
-        LOG.info("[AUDIT] Deleting service account key: {}", p12KeyName);
-        tryDeleteServiceAccountKey(p12KeyName);
-
-        LOG.info("[AUDIT] Deleting service account {} secret {}", serviceAcount, name);
-        client.secrets().delete(secret);
-      } catch (IOException | KubernetesClientException e) {
-        LOG.warn("[AUDIT] Failed to delete service account {} keys and/or secret {}",
-            serviceAcount, name);
-      }
-    }
+    serviceAccountSecretManager.cleanup();
   }
 
-  /**
-   * Try to delete a service account key, giving up with a warning if permission was denied.
-   * @param keyName The fully qualified name of the key to delete.
-   */
-  private void tryDeleteServiceAccountKey(String keyName) throws IOException {
-    try {
-      serviceAccountKeyManager.deleteKey(keyName);
-    } catch (GoogleJsonResponseException e) {
-      if (GcpUtil.isPermissionDenied(e)) {
-        LOG.warn("[AUDIT] Permission denied when trying to delete unused service account key {}");
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  private void ensureSecrets(WorkflowInstance workflowInstance, RunSpec runSpec, String secretEpoch)
+  private KubernetesSecretSpec ensureSecrets(WorkflowInstance workflowInstance, RunSpec runSpec)
       throws IOException {
-    ensureServiceAccountKeySecret(workflowInstance, runSpec, secretEpoch);
-    ensureCustomSecret(workflowInstance, runSpec);
+    return KubernetesSecretSpec.builder()
+        .customSecret(ensureCustomSecret(workflowInstance, runSpec))
+        .serviceAccountSecret(runSpec.serviceAccount().isPresent()
+            ? Optional.of(serviceAccountSecretManager.ensureServiceAccountKeySecret(
+                workflowInstance.workflowId().toString(), runSpec.serviceAccount().get()))
+            : Optional.empty())
+        .build();
   }
 
-  private void ensureCustomSecret(WorkflowInstance workflowInstance, RunSpec runSpec) {
-    runSpec.secret().ifPresent(specSecret -> {
+  private Optional<WorkflowConfiguration.Secret> ensureCustomSecret(
+      WorkflowInstance workflowInstance, RunSpec runSpec) {
+    return runSpec.secret().map(specSecret -> {
       if (specSecret.name().startsWith(STYX_WORKFLOW_SA_SECRET_NAME)) {
         LOG.warn("[AUDIT] Workflow {} refers to secret {} with managed service account key secret name prefix, "
             + "denying execution", specSecret.name());
@@ -277,108 +191,13 @@ class KubernetesDockerRunner implements DockerRunner {
         LOG.info("[AUDIT] Workflow {} refers to secret {}",
                  workflowInstance.workflowId(), specSecret.name());
       }
+
+      return specSecret;
     });
   }
 
-  private void ensureServiceAccountKeySecret(WorkflowInstance workflowInstance, RunSpec runSpec,
-      String secretEpoch) throws IOException {
-    if (!runSpec.serviceAccount().isPresent()) {
-      return;
-    }
-
-    final String serviceAccount = runSpec.serviceAccount().get();
-
-    // Check that the service account exists
-    final boolean serviceAccountExists = serviceAccountKeyManager.serviceAccountExists(serviceAccount);
-    if (!serviceAccountExists) {
-      LOG.error("[AUDIT] Workflow {} refers to non-existent service account {}", workflowInstance.workflowId(),
-          serviceAccount);
-      throw new InvalidExecutionException("Referenced service account " + serviceAccount + " was not found");
-    }
-
-    final String secretName = buildSecretName(serviceAccount, secretEpoch);
-
-    LOG.info("[AUDIT] Workflow {} refers to secret {} storing keys of {}",
-             workflowInstance.workflowId(), secretName, serviceAccount);
-
-    synchronized (secretMutationLock) {
-
-      // Check if we have a valid service account key secret already
-      final Secret existingSecret = client.secrets().withName(secretName).get();
-      if (existingSecret != null) {
-        if (serviceAccountKeysExist(existingSecret)) {
-          return;
-        }
-
-        LOG.info("[AUDIT] Service account keys have been deleted for {}, recreating",
-            serviceAccount);
-
-        // Need to delete this secret before creating a new one
-        client.secrets().delete(existingSecret);
-      }
-
-      // Create service account keys and secret
-      final ServiceAccountKey jsonKey;
-      final ServiceAccountKey p12Key;
-      try {
-        jsonKey = serviceAccountKeyManager.createJsonKey(serviceAccount);
-        p12Key = serviceAccountKeyManager.createP12Key(serviceAccount);
-      } catch (IOException e) {
-        LOG.error("[AUDIT] Failed to create keys for {}", serviceAccount, e);
-        throw e;
-      }
-
-      final Map<String, String> keys = ImmutableMap.of(
-          STYX_WORKFLOW_SA_JSON_KEY, jsonKey.getPrivateKeyData(),
-          STYX_WORKFLOW_SA_P12_KEY, p12Key.getPrivateKeyData()
-      );
-
-      final Map<String, String> annotations = ImmutableMap.of(
-          STYX_WORKFLOW_SA_JSON_KEY_NAME_ANNOTATION, jsonKey.getName(),
-          STYX_WORKFLOW_SA_P12_KEY_NAME_ANNOTATION, p12Key.getName(),
-          STYX_WORKFLOW_SA_ID_ANNOTATION, serviceAccount,
-          STYX_WORKFLOW_SA_EPOCH_ANNOTATION, secretEpoch
-      );
-
-      final Secret newSecret = new SecretBuilder()
-          .withNewMetadata()
-          .withName(secretName)
-          .withAnnotations(annotations)
-          .endMetadata()
-          .withData(keys)
-          .build();
-
-      client.secrets().create(newSecret);
-
-      LOG.info("[AUDIT] Secret {} created to store keys of {} referred by workflow {}",
-          secretName, serviceAccount, workflowInstance.workflowId());
-    }
-  }
-
-  private boolean serviceAccountKeysExist(Secret secret) {
-    final Map<String, String> annotations = secret.getMetadata().getAnnotations();
-    final String jsonKeyName = annotations.get(STYX_WORKFLOW_SA_JSON_KEY_NAME_ANNOTATION);
-    final String p12KeyName = annotations.get(STYX_WORKFLOW_SA_P12_KEY_NAME_ANNOTATION);
-
-    try {
-      return serviceAccountKeyManager.keyExists(jsonKeyName)
-          && serviceAccountKeyManager.keyExists(p12KeyName);
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private static String buildSecretName(String serviceAccount, String epoch) {
-    return STYX_WORKFLOW_SA_SECRET_NAME + '-' + epoch + '-'
-           + Hashing.sha256().hashString(serviceAccount, UTF_8);
-  }
-
-  private static String utcNoonSecretEpoch() {
-    return Long.toString((System.currentTimeMillis() + TimeUnit.HOURS.toMillis(12)) / TimeUnit.HOURS.toMillis(24));
-  }
-
   @VisibleForTesting
-  static Pod createPod(WorkflowInstance workflowInstance, RunSpec runSpec, String epoch) {
+  static Pod createPod(WorkflowInstance workflowInstance, RunSpec runSpec, KubernetesSecretSpec secretSpec) {
     final String imageWithTag = runSpec.imageName().contains(":")
         ? runSpec.imageName()
         : runSpec.imageName() + ":latest";
@@ -402,9 +221,9 @@ class KubernetesDockerRunner implements DockerRunner {
         .withArgs(runSpec.args())
         .withEnv(buildEnv(workflowInstance, runSpec, podName));
 
-    if (runSpec.serviceAccount().isPresent()) {
+    secretSpec.serviceAccountSecret().ifPresent(serviceAccountSecret -> {
       final SecretVolumeSource saVolumeSource = new SecretVolumeSourceBuilder()
-          .withSecretName(buildSecretName(runSpec.serviceAccount().get(), epoch))
+          .withSecretName(serviceAccountSecret)
           .build();
       final Volume saVolume = new VolumeBuilder()
           .withName(STYX_WORKFLOW_SA_SECRET_NAME)
@@ -421,10 +240,9 @@ class KubernetesDockerRunner implements DockerRunner {
       // TODO: do we need set this env as default value?
       containerBuilder.addToEnv(envVar(STYX_WORKFLOW_SA_ENV_VARIABLE,
                                        saMount.getMountPath() + STYX_WORKFLOW_SA_JSON_KEY));
-    }
+    });
 
-    if (runSpec.secret().isPresent()) {
-      final WorkflowConfiguration.Secret secret = runSpec.secret().get();
+    secretSpec.customSecret().ifPresent(secret -> {
       final SecretVolumeSource secretVolumeSource = new SecretVolumeSourceBuilder()
           .withSecretName(secret.name())
           .build();
@@ -440,7 +258,8 @@ class KubernetesDockerRunner implements DockerRunner {
           .withReadOnly(true)
           .build();
       containerBuilder.addToVolumeMounts(secretMount);
-    }
+    });
+
     specBuilder.addToContainers(containerBuilder.build());
     podBuilder.withSpec(specBuilder.build());
 
@@ -758,6 +577,16 @@ class KubernetesDockerRunner implements DockerRunner {
     @Override
     public Boolean halt(WorkflowInstance workflowInstance) {
       return false;
+    }
+  }
+
+  @AutoMatter
+  interface KubernetesSecretSpec {
+    Optional<WorkflowConfiguration.Secret> customSecret();
+    Optional<String> serviceAccountSecret();
+
+    static KubernetesSecretSpecBuilder builder() {
+      return new KubernetesSecretSpecBuilder();
     }
   }
 }
