+++ styx-scheduler-service/src/test/java/com/spotify/styx/docker/KubernetesDockerRunnerTest.java
@@ -25,16 +25,12 @@ import static com.spotify.styx.docker.KubernetesPodEventTranslatorTest.podStatus
 import static com.spotify.styx.docker.KubernetesPodEventTranslatorTest.running;
 import static com.spotify.styx.docker.KubernetesPodEventTranslatorTest.terminated;
 import static com.spotify.styx.docker.KubernetesPodEventTranslatorTest.waiting;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Optional.empty;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.awaitility.Awaitility.await;
-import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.times;
@@ -42,38 +38,29 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
-import com.google.api.client.googleapis.json.GoogleJsonError;
-import com.google.api.client.googleapis.json.GoogleJsonResponseException;
-import com.google.api.client.http.HttpHeaders;
-import com.google.api.client.http.HttpResponseException.Builder;
-import com.google.api.services.iam.v1.model.ServiceAccountKey;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.hash.Hashing;
-import com.spotify.styx.ServiceAccountKeyManager;
 import com.spotify.styx.docker.DockerRunner.RunSpec;
+import com.spotify.styx.docker.KubernetesDockerRunner.KubernetesSecretSpec;
 import com.spotify.styx.model.Event;
 import com.spotify.styx.model.WorkflowConfiguration;
 import com.spotify.styx.model.WorkflowInstance;
 import com.spotify.styx.monitoring.Stats;
-import com.spotify.styx.serialization.Json;
 import com.spotify.styx.state.RunState;
 import com.spotify.styx.state.StateData;
 import com.spotify.styx.state.StateManager;
-import com.spotify.styx.state.StateManager.IsClosed;
 import com.spotify.styx.state.SyncStateManager;
 import com.spotify.styx.testdata.TestData;
 import io.fabric8.kubernetes.api.model.DoneablePod;
 import io.fabric8.kubernetes.api.model.DoneableSecret;
 import io.fabric8.kubernetes.api.model.ListMeta;
-import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
 import io.fabric8.kubernetes.api.model.PodList;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.SecretBuilder;
 import io.fabric8.kubernetes.api.model.SecretList;
-import io.fabric8.kubernetes.client.KubernetesClientException;
+import io.fabric8.kubernetes.api.model.SecretVolumeSource;
+import io.fabric8.kubernetes.api.model.Volume;
 import io.fabric8.kubernetes.client.NamespacedKubernetesClient;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
@@ -82,12 +69,9 @@ import io.fabric8.kubernetes.client.dsl.PodResource;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.Watchable;
 import java.io.IOException;
-import java.time.Clock;
-import java.time.Duration;
-import java.time.Instant;
-import java.time.ZoneOffset;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import org.junit.After;
 import org.junit.Before;
@@ -106,6 +90,7 @@ public class KubernetesDockerRunnerTest {
 
   private static final String POD_NAME = "test-pod-1";
   private static final String SERVICE_ACCOUNT = "sa@example.com";
+  private static final String SERVICE_ACCOUNT_SECRET = "sa-secret";
   private static final WorkflowInstance WORKFLOW_INSTANCE = WorkflowInstance.create(TestData.WORKFLOW_ID, "foo");
   private static final RunSpec RUN_SPEC = RunSpec.simple("busybox");
   private static final RunSpec RUN_SPEC_WITH_SECRET = RunSpec.create("busybox",
@@ -120,6 +105,18 @@ public class KubernetesDockerRunnerTest {
                                                                  empty(),
                                                                  Optional.of(SERVICE_ACCOUNT),
                                                                  empty());
+
+  private static final KubernetesSecretSpec SECRET_SPEC_WITH_SA = KubernetesSecretSpec.builder()
+      .serviceAccountSecret(SERVICE_ACCOUNT_SECRET)
+      .build();
+
+  private static final KubernetesSecretSpec SECRET_SPEC_WITH_CUSTOM_SECRET = KubernetesSecretSpec.builder()
+      .customSecret(RUN_SPEC_WITH_SECRET.secret())
+      .build();
+
+  private static final KubernetesSecretSpec EMPTY_SECRET_SPEC = KubernetesSecretSpec.builder()
+      .build();
+
   private static final RunSpec RUN_SPEC_WITH_SECRET_AND_SA = RunSpec.create("busybox",
                                                                             ImmutableList.copyOf(new String[0]),
                                                                             false,
@@ -127,33 +124,24 @@ public class KubernetesDockerRunnerTest {
                                                                             Optional.of(SERVICE_ACCOUNT),
                                                                             empty());
 
-  private final static String SECRET_EPOCH = "4711";
-
-  private final static Clock CLOCK = Clock.fixed(Instant.now(), ZoneOffset.UTC);
-
   @Mock NamespacedKubernetesClient k8sClient;
 
-  @Mock ServiceAccountKeyManager serviceAccountKeyManager;
+  @Mock KubernetesGCPServiceAccountSecretManager serviceAccountSecretManager;
 
   @Mock MixedOperation<Pod, PodList, DoneablePod, PodResource<Pod, DoneablePod>> pods;
   @Mock MixedOperation<Secret, SecretList, DoneableSecret, Resource<Secret, DoneableSecret>> secrets;
   @Mock Resource<Secret, DoneableSecret> namedResource;
-  @Mock SecretList secretList;
 
   @Mock PodList podList;
   @Mock ListMeta listMeta;
   @Mock Watchable<Watch, Watcher<Pod>> podWatchable;
   @Mock Watch watch;
   @Captor ArgumentCaptor<Watcher<Pod>> watchCaptor;
-  @Captor ArgumentCaptor<Secret> secretCaptor;
+  @Captor ArgumentCaptor<Pod> podCaptor;
 
-  @Rule
-  public ExpectedException exception = ExpectedException.none();
+  @Rule public ExpectedException exception = ExpectedException.none();
 
-  ServiceAccountKey serviceAccountJsonKey = new ServiceAccountKey();
-  ServiceAccountKey serviceAccountP12Key = new ServiceAccountKey();
-
-  Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC, SECRET_EPOCH);
+  Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC, EMPTY_SECRET_SPEC);
   StateManager stateManager = Mockito.spy(new SyncStateManager());
   Stats stats = Mockito.mock(Stats.class);
 
@@ -174,11 +162,11 @@ public class KubernetesDockerRunnerTest {
     when(pods.withResourceVersion("1000")).thenReturn(podWatchable);
     when(podWatchable.watch(watchCaptor.capture())).thenReturn(watch);
 
-    when(serviceAccountKeyManager.createJsonKey(anyString())).thenReturn(serviceAccountJsonKey);
-    when(serviceAccountKeyManager.createJsonKey(anyString())).thenReturn(serviceAccountP12Key);
+    when(serviceAccountSecretManager.ensureServiceAccountKeySecret(
+        WORKFLOW_INSTANCE.workflowId().toString(), SERVICE_ACCOUNT))
+        .thenReturn(SERVICE_ACCOUNT_SECRET);
 
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager, 60, () -> SECRET_EPOCH,
-        CLOCK);
+    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountSecretManager, 60);
     kdr.init();
     kdr.restore();
 
@@ -193,10 +181,7 @@ public class KubernetesDockerRunnerTest {
     StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
     stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
 
-    when(pods.create(any(Pod.class))).thenReturn(createdPod);
-
-    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC);
-    stateManager.receive(Event.started(WORKFLOW_INSTANCE));
+    when(pods.create(podCaptor.capture())).thenReturn(createdPod);
   }
 
   @After
@@ -209,16 +194,10 @@ public class KubernetesDockerRunnerTest {
     when(secrets.withName(any(String.class))).thenReturn(namedResource);
     when(namedResource.get()).thenReturn(null);
     when(k8sClient.secrets()).thenReturn(secrets);
-    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET, SECRET_EPOCH);
+    Pod createdPod = KubernetesDockerRunner.createPod(
+        WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET, SECRET_SPEC_WITH_CUSTOM_SECRET);
     when(pods.create(any(Pod.class))).thenReturn(createdPod);
 
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    kdr.init();
     kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET);
   }
 
@@ -227,22 +206,17 @@ public class KubernetesDockerRunnerTest {
     when(secrets.withName(any(String.class))).thenReturn(namedResource);
     when(namedResource.get()).thenReturn(null);
     when(k8sClient.secrets()).thenReturn(secrets);
-    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET_AND_SA, SECRET_EPOCH);
+    Pod createdPod = KubernetesDockerRunner.createPod(
+        WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET_AND_SA, SECRET_SPEC_WITH_CUSTOM_SECRET);
     when(pods.create(any(Pod.class))).thenReturn(createdPod);
 
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    kdr.init();
     kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET_AND_SA);
   }
 
   @Test
   public void shouldMountSecret() throws IOException, StateManager.IsClosed {
-    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET, SECRET_EPOCH);
+    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET,
+        SECRET_SPEC_WITH_CUSTOM_SECRET);
     assertThat(pod.getSpec().getVolumes().size(), is(1));
     assertThat(pod.getSpec().getVolumes().get(0).getName(),
                is(RUN_SPEC_WITH_SECRET.secret().get().name()));
@@ -254,7 +228,7 @@ public class KubernetesDockerRunnerTest {
 
   @Test
   public void shouldMountServiceAccount() throws IOException, StateManager.IsClosed {
-    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
+    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_SPEC_WITH_SA);
     assertThat(pod.getSpec().getVolumes().size(), is(1));
     assertThat(pod.getSpec().getVolumes().get(0).getName(),
                is(KubernetesDockerRunner.STYX_WORKFLOW_SA_SECRET_NAME));
@@ -271,351 +245,67 @@ public class KubernetesDockerRunnerTest {
     when(secrets.withName(any(String.class))).thenReturn(namedResource);
     when(namedResource.get()).thenReturn(new SecretBuilder().build());
     when(k8sClient.secrets()).thenReturn(secrets);
-    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET, SECRET_EPOCH);
+    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET,
+        SECRET_SPEC_WITH_CUSTOM_SECRET);
     when(pods.create(any(Pod.class))).thenReturn(createdPod);
 
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
-    stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
-    kdr.init();
     kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SECRET);
-    stateManager.receive(Event.started(WORKFLOW_INSTANCE));
+
+    verify(pods).create(podCaptor.capture());
   }
 
   @Test
   public void shouldCleanupServiceAccountSecrets() throws Exception {
-    final String creationTimestamp = CLOCK.instant().minus(Duration.ofHours(1)).toString();
-    final Secret secret = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, SECRET_EPOCH, "json-key", "p12-key", creationTimestamp);
-
-    when(k8sClient.secrets()).thenReturn(secrets);
-    when(secrets.list()).thenReturn(secretList);
-    when(secretList.getItems()).thenReturn(ImmutableList.of(secret));
-
-    // Verify that a service account key secret in use is not deleted
-    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
-    when(podList.getItems()).thenReturn(ImmutableList.of(pod));
     kdr.cleanup();
-    verify(serviceAccountKeyManager, never()).deleteKey(anyString());
-    verify(secrets, never()).delete(any(Secret.class));
-
-    // Verify that an unused service account key secret is deleted
-    when(podList.getItems()).thenReturn(ImmutableList.of());
-    kdr.cleanup();
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "json-key"));
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "p12-key"));
-    verify(secrets).delete(secret);
+    verify(serviceAccountSecretManager).cleanup();
   }
 
   @Test
-  public void shouldRemoveServiceAccountSecretsInPastEpoch() throws Exception {
-    final String creationTimestamp = CLOCK.instant().minus(Duration.ofHours(1)).toString();
-    final Secret secret = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, "pre-" + SECRET_EPOCH, "old-json-key", "old-p12-key", creationTimestamp);
-
-    when(k8sClient.secrets()).thenReturn(secrets);
-    when(secrets.list()).thenReturn(secretList);
-    when(secretList.getItems()).thenReturn(ImmutableList.of(secret));
-
-    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
-    when(podList.getItems()).thenReturn(ImmutableList.of(pod));
-    kdr.cleanup();
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "old-json-key"));
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "old-p12-key"));
-    verify(secrets).delete(secret);
-  }
-
-  @Test
-  public void shouldNotRemoveRecentlyCreatedServiceAccountSecrets() throws Exception {
-    final String creationTimestamp = CLOCK.instant().toString();
-    final Secret secret1 = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, "pre-" + SECRET_EPOCH, "json-key-1", "p12-key-1", creationTimestamp);
-    final Secret secret2 = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, SECRET_EPOCH, "json-key-2", "p12-key-2", creationTimestamp);
-
-    when(k8sClient.secrets()).thenReturn(secrets);
-    when(secrets.list()).thenReturn(secretList);
-    when(secretList.getItems()).thenReturn(ImmutableList.of(secret1, secret2));
-
-    final Pod pod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
-    when(podList.getItems()).thenReturn(ImmutableList.of(pod));
-    kdr.cleanup();
-    verify(serviceAccountKeyManager, never()).deleteKey(anyString());
-    verify(secrets, never()).delete(any(Secret.class));
-  }
-
-  @Test
-  public void shouldHandlePermissionDeniedErrorsWhenDeletingServiceAccountKeys() throws Exception {
-    final String creationTimestamp = CLOCK.instant().minus(Duration.ofHours(1)).toString();
-    final Secret secret = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, SECRET_EPOCH, "json-key", "p12-key", creationTimestamp);
-
-    when(podList.getItems()).thenReturn(ImmutableList.of());
-    when(k8sClient.secrets()).thenReturn(secrets);
-    when(secrets.list()).thenReturn(secretList);
-    when(secretList.getItems()).thenReturn(ImmutableList.of(secret));
-
-    // Verify that the secret is delete even if we get permission denied errors on deleting the keys
-    final GoogleJsonResponseException permissionDenied = new GoogleJsonResponseException(
-        new Builder(403, "Forbidden", new HttpHeaders()), new GoogleJsonError()
-        .set("status", "PERMISSION_DENIED"));
-    doThrow(permissionDenied).when(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "json-key"));
-    doThrow(permissionDenied).when(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "p12-key"));
-    kdr.cleanup();
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "json-key"));
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "p12-key"));
-    verify(secrets).delete(secret);
-  }
-
-  @Test
-  public void shouldHandleErrorsWhenDeletingServiceAccountKeys() throws Exception {
-    final String creationTimestamp = CLOCK.instant().minus(Duration.ofHours(1)).toString();
-    final Secret secret1 = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, SECRET_EPOCH, "json-key-1", "p12-key-1", creationTimestamp);
-    final Secret secret2 = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, SECRET_EPOCH, "json-key-2", "p12-key-2", creationTimestamp);
-    final Secret secret3 = fakeServiceAccountKeySecret(
-        SERVICE_ACCOUNT, SECRET_EPOCH, "json-key-3", "p12-key-3", creationTimestamp);
-
-    when(podList.getItems()).thenReturn(ImmutableList.of());
-    when(k8sClient.secrets()).thenReturn(secrets);
-    when(secrets.list()).thenReturn(secretList);
-    when(secretList.getItems()).thenReturn(ImmutableList.of(secret1, secret2, secret3));
-
-    when(secrets.delete(secret1)).thenThrow(new KubernetesClientException("fail delete secret1"));
-    doThrow(new IOException("fail delete json-key-2")).when(serviceAccountKeyManager).deleteKey("json-key-2");
-    doThrow(new IOException("fail delete p12-key-2")).when(serviceAccountKeyManager).deleteKey("p12-key-2");
-
-    kdr.cleanup();
-
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "json-key-1"));
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "p12-key-1"));
-    verify(secrets).delete(secret1);
-
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "json-key-2"));
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "p12-key-2"));
-    verify(secrets).delete(secret2);
-
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "json-key-3"));
-    verify(serviceAccountKeyManager).deleteKey(keyName(SERVICE_ACCOUNT, "p12-key-3"));
-    verify(secrets).delete(secret3);
-  }
-
-  private static Secret fakeServiceAccountKeySecret(String serviceAccount, String epoch, String jsonKeyId,
-      String p12KeyId, String creationTimestamp) {
-    final String jsonKeyName = keyName(serviceAccount, jsonKeyId);
-    final String p12KeyName = keyName(serviceAccount, p12KeyId);
-
-    final ObjectMeta metadata = new ObjectMeta();
-    metadata.setCreationTimestamp(creationTimestamp);
-    metadata.setName("styx-wf-sa-keys-" + epoch + "-" + Hashing.sha256().hashString(serviceAccount, UTF_8));
-    metadata.setAnnotations(ImmutableMap.of(
-        "styx-wf-sa", serviceAccount,
-        "styx-wf-sa-json-key-name", jsonKeyName,
-        "styx-wf-sa-p12-key-name", p12KeyName));
-
-    return new SecretBuilder()
-        .withMetadata(metadata)
-        .withData(ImmutableMap.of(
-            "styx-wf-sa.json", "json-private-key-data",
-            "styx-wf-sa.p12", "p12-private-key-data"))
-        .build();
-  }
-
-  @Test
-  public void shouldCreateSASecret() throws StateManager.IsClosed, IOException {
+  public void shouldEnsureAndMountServiceAccountSecret() throws StateManager.IsClosed, IOException {
     when(secrets.withName(any(String.class))).thenReturn(namedResource);
     when(namedResource.get()).thenReturn(null);
     when(k8sClient.secrets()).thenReturn(secrets);
 
-    when(serviceAccountKeyManager.serviceAccountExists(SERVICE_ACCOUNT)).thenReturn(true);
-
-    ServiceAccountKey jsonKey = new ServiceAccountKey();
-    jsonKey.setName("key.json");
-    jsonKey.setPrivateKeyData("json-private-key-data");
-    ServiceAccountKey p12Key = new ServiceAccountKey();
-    p12Key.setName("key.p12");
-    p12Key.setPrivateKeyData("p12-private-key-data");
-    when(serviceAccountKeyManager.createJsonKey(any(String.class))).thenReturn(jsonKey);
-    when(serviceAccountKeyManager.createP12Key(any(String.class))).thenReturn(p12Key);
-
-    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
+    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_SPEC_WITH_SA);
     when(pods.create(any(Pod.class))).thenReturn(createdPod);
 
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
+    when(serviceAccountSecretManager.ensureServiceAccountKeySecret(
+        WORKFLOW_INSTANCE.workflowId().toString(), SERVICE_ACCOUNT)).thenReturn(SERVICE_ACCOUNT_SECRET);
 
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
     StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
     stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
-    kdr.init();
-    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA);
-    verify(secrets).create(secretCaptor.capture());
 
-    final Secret createdSecret = secretCaptor.getValue();
-    assertThat(createdSecret.getMetadata().getAnnotations(), hasEntry("styx-wf-sa", SERVICE_ACCOUNT));
-    assertThat(createdSecret.getData(), hasEntry("styx-wf-sa.json", jsonKey.getPrivateKeyData()));
-    assertThat(createdSecret.getData(), hasEntry("styx-wf-sa.p12", p12Key.getPrivateKeyData()));
+    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA);
+
+    verify(serviceAccountSecretManager).ensureServiceAccountKeySecret(
+        WORKFLOW_INSTANCE.workflowId().toString(), SERVICE_ACCOUNT);
+
+    verify(pods).create(podCaptor.capture());
+
+    final Pod pod = podCaptor.getValue();
+
+    final Optional<SecretVolumeSource> serviceAccountSecretVolume = pod.getSpec().getVolumes().stream()
+        .map(Volume::getSecret)
+        .filter(Objects::nonNull)
+        .filter(v -> SERVICE_ACCOUNT_SECRET.equals(v.getSecretName()))
+        .findAny();
+
+    assertThat(serviceAccountSecretVolume.isPresent(), is(true));
   }
 
   @Test
-  public void shouldRunIfSASecretExists() throws StateManager.IsClosed, IOException {
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
+  public void shouldNotRunIfServiceAccountSecretEnsureFails() throws StateManager.IsClosed, IOException {
+    final InvalidExecutionException error = new InvalidExecutionException("SA not found");
+    when(serviceAccountSecretManager.ensureServiceAccountKeySecret(
+        WORKFLOW_INSTANCE.workflowId().toString(), SERVICE_ACCOUNT)).thenThrow(error);
 
-    final String jsonKeyId = "json-key";
-    final String p12KeyId = "p12-key";
-
-    final String jsonKeyName = keyName(SERVICE_ACCOUNT, jsonKeyId);
-    final String p12KeyName = keyName(SERVICE_ACCOUNT, p12KeyId);
-
-    final String jsonKeyData = "json-private-key-data";
-    final String p12KeyData = "p12-private-key-data";
-
-    final ObjectMeta metadata = new ObjectMeta();
-    metadata.setAnnotations(ImmutableMap.of(
-        "styx-wf-sa", SERVICE_ACCOUNT,
-        "styx-wf-sa-json-key-name", jsonKeyName,
-        "styx-wf-sa-p12-key-name", p12KeyName));
-
-    final Secret secret = new SecretBuilder()
-        .withMetadata(metadata)
-        .withData(ImmutableMap.of(
-            "styx-wf-sa.json", jsonKeyData,
-            "styx-wf-sa.p12", p12KeyData))
-        .build();
-
-    when(serviceAccountKeyManager.serviceAccountExists(SERVICE_ACCOUNT)).thenReturn(true);
-    when(serviceAccountKeyManager.keyExists(jsonKeyName)).thenReturn(true);
-    when(serviceAccountKeyManager.keyExists(p12KeyName)).thenReturn(true);
-
-    when(secrets.withName(any(String.class))).thenReturn(namedResource);
-    when(namedResource.get()).thenReturn(secret);
-
-    when(k8sClient.secrets()).thenReturn(secrets);
-    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
-    when(pods.create(any(Pod.class))).thenReturn(createdPod);
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
-    stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
-    kdr.init();
-    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA);
-
-    verify(secrets, never()).create(any(Secret.class));
-  }
-
-  private static String keyName(String serviceAccount, String keyId) {
-    return "projects/-/serviceAccounts/" + serviceAccount + "/keys/" + keyId;
-  }
-
-  @Test
-  public void shouldNotRunIfServiceAccountDoesNotExist() throws StateManager.IsClosed, IOException {
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
-
-    when(serviceAccountKeyManager.serviceAccountExists(SERVICE_ACCOUNT)).thenReturn(false);
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
-    stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
-    kdr.init();
-
-    exception.expect(InvalidExecutionException.class);
-    exception.expectMessage("Referenced service account " + SERVICE_ACCOUNT + " was not found");
+    exception.expect(is(error));
 
     kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA);
-  }
-
-  @Test
-  public void shouldCreateNewServiceAccountKeysIfKeysAreDeleted() throws IOException, IsClosed {
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
-
-    final ObjectMeta metadata = new ObjectMeta();
-    metadata.setAnnotations(ImmutableMap.of("styx-wf-sa", SERVICE_ACCOUNT));
-
-    final String jsonKeyName = "json-key";
-    final String p12KeyName = "p12-key";
-    final String newJsonKeyName = "new-json-key";
-    final String newP12KeyName = "new-p12-key";
-
-    final ServiceAccountKey jsonKey = new ServiceAccountKey()
-        .setName(jsonKeyName)
-        .setPrivateKeyData("json-private-key-data");
-
-    final ServiceAccountKey p12Key = new ServiceAccountKey()
-        .setName(p12KeyName)
-        .setPrivateKeyData("p12-private-key-data");
-
-    final ServiceAccountKey newJsonKey = new ServiceAccountKey()
-        .setName(newJsonKeyName)
-        .setPrivateKeyData("new-json-private-key-data");
-
-    final ServiceAccountKey newP12Key = new ServiceAccountKey()
-        .setName(newP12KeyName)
-        .setPrivateKeyData("new-p12-private-key-data");
-
-
-    when(serviceAccountKeyManager.serviceAccountExists(SERVICE_ACCOUNT)).thenReturn(true);
-    when(serviceAccountKeyManager.keyExists(keyName(SERVICE_ACCOUNT, jsonKeyName))).thenReturn(false);
-    when(serviceAccountKeyManager.keyExists(keyName(SERVICE_ACCOUNT, p12KeyName))).thenReturn(false);
-
-    when(serviceAccountKeyManager.createJsonKey(any(String.class))).thenReturn(newJsonKey);
-    when(serviceAccountKeyManager.createP12Key(any(String.class))).thenReturn(newP12Key);
-
-    when(secrets.withName(any(String.class))).thenReturn(namedResource);
-    when(namedResource.get()).thenReturn(new SecretBuilder()
-        .withMetadata(metadata)
-        .withData(ImmutableMap.of(
-            "styx-wf-sa.json", Json.OBJECT_MAPPER.writeValueAsString(jsonKey),
-            "styx-wf-sa.p12", Json.OBJECT_MAPPER.writeValueAsString(p12Key)))
-        .build());
-
-    when(k8sClient.secrets()).thenReturn(secrets);
-    Pod createdPod = KubernetesDockerRunner.createPod(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA, SECRET_EPOCH);
-    when(pods.create(any(Pod.class))).thenReturn(createdPod);
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
-    stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
-    kdr.init();
-
-    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC_WITH_SA);
-
-    verify(secrets).create(secretCaptor.capture());
-
-    final Secret createdSecret = secretCaptor.getValue();
-    assertThat(createdSecret.getMetadata().getAnnotations(), hasEntry("styx-wf-sa", SERVICE_ACCOUNT));
-    assertThat(createdSecret.getData(), hasEntry("styx-wf-sa.json", newJsonKey.getPrivateKeyData()));
-    assertThat(createdSecret.getData(), hasEntry("styx-wf-sa.p12", newP12Key.getPrivateKeyData()));
   }
 
   @Test
   public void shouldNotRunIfSecretHasManagedServiceAccountKeySecretNamePrefix() throws StateManager.IsClosed, IOException {
-    stateManager.receive(Event.terminate(WORKFLOW_INSTANCE, Optional.of(0)));
-    stateManager.receive(Event.success(WORKFLOW_INSTANCE));
-    kdr.close();
-
-    // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
-    StateData stateData = StateData.newBuilder().executionId(POD_NAME).build();
-    stateManager.initialize(RunState.create(WORKFLOW_INSTANCE, RunState.State.SUBMITTED, stateData));
-    kdr.init();
-
     final String secret = "styx-wf-sa-keys-foo";
 
     exception.expect(InvalidExecutionException.class);
@@ -674,6 +364,9 @@ public class KubernetesDockerRunnerTest {
 
   @Test
   public void shouldIgnoreDeletedEvents() throws Exception {
+    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC);
+    stateManager.receive(Event.started(WORKFLOW_INSTANCE));
+
     podWatcher.eventReceived(Watcher.Action.DELETED, createdPod);
 
     assertThat(stateManager.get(WORKFLOW_INSTANCE).state(), is(RunState.State.RUNNING));
@@ -711,6 +404,9 @@ public class KubernetesDockerRunnerTest {
 
   @Test
   public void shouldDiscardChangesForOldExecutions() throws Exception {
+    kdr.start(WORKFLOW_INSTANCE, RUN_SPEC);
+    stateManager.receive(Event.started(WORKFLOW_INSTANCE));
+
     // simulate event from different pod, but still with the same workflow instance annotation
     createdPod.getMetadata().setName(POD_NAME + "-other");
     createdPod.setStatus(terminated("Succeeded", 20, null));
@@ -727,7 +423,7 @@ public class KubernetesDockerRunnerTest {
     createdPod.setStatus(terminated("Succeeded", 20, null));
 
     // Start a new runner
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager);
+    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountSecretManager);
     kdr.init();
 
     // Make the runner poll states for all pods
@@ -744,8 +440,7 @@ public class KubernetesDockerRunnerTest {
 
     // Set up a runner with short poll interval to avoid this test having to wait a long time for the poll
     kdr.close();
-    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountKeyManager, 1, () -> SECRET_EPOCH,
-        CLOCK);
+    kdr = new KubernetesDockerRunner(k8sClient, stateManager, stats, serviceAccountSecretManager, 1);
     kdr.init();
     kdr.restore();
